re-structured free apis file
