+++ src/common/interfaces/requestProps.interface.ts
@@ -1,0 +1,9 @@
+export interface requestProp {
+  payload: any;
+  apiId: string;
+  method: string;
+  endpoint: string;
+  base_url: string;
+  profileId: string;
+  secretKey: string;
+}
