+++ src/subscription/apis.ts
@@ -1,92 +1,22 @@
 export const FreeApis = [
   {
     id: 'e986680a-128f-44c6-bb14-b20bb134ec6e',
-    createdOn: '2022-10-13T14:38:44.859Z',
-    createdBy: null,
-    updatedOn: null,
-    updatedBy: null,
-    deletedOn: null,
-    deletedBy: null,
     name: 'Text Summerizer',
-    description: 'Summarizes a block of text passed to it',
     base_url:
       'https://summaryer-api.pk25mf6178910.eu-west-3.cs.amazonlightsail.com',
-    popularity: 0,
-    about: null,
-    logo_url: null,
-    api_website: null,
-    term_of_use: null,
-    read_me: null,
-    subscriptions: [],
-    status: 'unverified',
-    visibility: 'private',
-    rating: 0,
-    service_level: 0,
-    latency: 0,
-    categoryId: '92275577-46d6-43c4-bb15-dc2fb4bb56f9',
+    route: '/text',
+    method: 'post',
     profileId: '922c7a82-d628-45e5-b338-cf3a44eb0834',
     secretKey: '78cfefa0-a100-4adb-bcb2-af8d6e237eeb',
-    tutorialsId: null,
-    endpoint: {
-      id: '17554211-c19c-4e2e-87f4-1f480b074202',
-      createdOn: '2022-10-13T14:57:31.622Z',
-      createdBy: null,
-      updatedOn: null,
-      updatedBy: null,
-      deletedOn: null,
-      deletedBy: null,
-      name: 'Text',
-      method: 'post',
-      route: '/text',
-      apiId: 'e986680a-128f-44c6-bb14-b20bb134ec6e',
-      description: 'Summarizes a block of text passed to it',
-      headers: [],
-      requestBody: [],
-    },
   },
   {
     id: 'fe254b53-16f2-4001-8eb2-3158b9d5840d',
-    createdOn: '2022-10-13T13:39:44.777Z',
-    createdBy: null,
-    updatedOn: null,
-    updatedBy: null,
-    deletedOn: null,
-    deletedBy: null,
     name: 'Question & Answer',
-    description: 'Ask a question based on a supplied context and get an answer',
     base_url:
       'https://qnanswer-api.pk25mf6178910.eu-west-3.cs.amazonlightsail.com',
-    popularity: 0,
-    about: null,
-    logo_url: null,
-    api_website: null,
-    term_of_use: null,
-    read_me: null,
-    subscriptions: [],
-    status: 'unverified',
-    visibility: 'private',
-    rating: 0,
-    service_level: 0,
-    latency: 0,
-    categoryId: '92275577-46d6-43c4-bb15-dc2fb4bb56f9',
+    method: 'post',
+    route: '/q_and_a',
     profileId: '922c7a82-d628-45e5-b338-cf3a44eb0834',
     secretKey: 'a2b91b77-b370-4e6c-8cab-d5d63cedf9b1',
-    tutorialsId: null,
-    endpoint: {
-      id: '75a4d917-0e42-49e1-8d59-697734e2ac10',
-      createdOn: '2022-10-13T14:10:08.958Z',
-      createdBy: null,
-      updatedOn: null,
-      updatedBy: null,
-      deletedOn: null,
-      deletedBy: null,
-      name: 'question & answer',
-      method: 'post',
-      route: '/q_and_a',
-      apiId: 'fe254b53-16f2-4001-8eb2-3158b9d5840d',
-      description: 'get answer based on a question and supplied context',
-      headers: [],
-      requestBody: [],
-    },
   },
 ];
