+++ src/subscription/httpCall.service.ts
@@ -1,12 +1,8 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
 import { ZaLaResponse } from '../common/helpers/response';
-import { Api } from '../entities/api.entity';
-import { Repository } from 'typeorm';
-import { Subscription } from '../entities/subscription.entity';
 import { HttpService } from '@nestjs/axios';
 import { AnalyticsService } from 'src/analytics/analytics.service';
-import { requestProp } from 'src/common/interfaces/freeApis.interface';
+import { requestProp } from 'src/common/interfaces/requestProps.interface';
 
 @Injectable()
 export class HttpCallService {
@@ -15,7 +11,15 @@ export class HttpCallService {
     private readonly analyticsService: AnalyticsService,
   ) {}
 
-  async call({ api, profileId, endpoint, payload }: requestProp): Promise<any> {
+  async call({
+    apiId,
+    base_url,
+    payload,
+    profileId,
+    endpoint,
+    secretKey,
+    method,
+  }: requestProp): Promise<any> {
     try {
       /* Getting the current time in nanoseconds. */
       const startTime = process.hrtime();
@@ -23,10 +27,10 @@ export class HttpCallService {
       /* Making a request to the api with the payload and the secret key. */
       const ref = this.httpService.axiosRef;
       const axiosResponse = await ref({
-        method: endpoint.method.toLowerCase(),
-        url: `${api.base_url}${endpoint.route}`,
+        method,
+        url: `${base_url}${endpoint}`,
         data: payload,
-        headers: { 'X-Zapi-Proxy-Secret': api.secretKey },
+        headers: { 'X-Zapi-Proxy-Secret': secretKey },
       });
 
       /* Calculating the time it takes to make a request to the api. */
@@ -37,28 +41,28 @@ export class HttpCallService {
 
       this.analyticsService.updateAnalytics(
         axiosResponse.status,
-        api.id,
+        apiId,
         totalTimeInMs,
       );
       this.analyticsService.analyticLogs({
         status: axiosResponse.status,
         latency: Math.round(totalTimeInMs),
         profileId,
-        apiId: api.id,
-        endpoint: endpoint.route,
-        method: endpoint.method.toLowerCase(),
+        apiId: apiId,
+        endpoint,
+        method,
       });
 
       return data;
     } catch (error) {
-      this.analyticsService.updateAnalytics(error.response.status, api.id);
+      this.analyticsService.updateAnalytics(error.response.status, apiId);
       this.analyticsService.analyticLogs({
         status: error.response.status,
         errorMessage: error.response.statusText,
         profileId,
-        apiId: api.id,
-        endpoint: endpoint.route,
-        method: endpoint.method.toLowerCase(),
+        apiId,
+        endpoint,
+        method,
       });
 
       throw new BadRequestException(
